neomutt: mpv integration: simplify and clean up
Just pipe the whole mail to mpvc-enqueue, and set pipe_decode. This way
neomutt will already the mail for us using the attachment preferences
from .mailcap, even rendering any HTML in text/html parts, if those are
shown, so it will pipe the same text to the external command like when
we would be viewing the mail in the pager. We also don't need to go to
the attachment menu, which results in less UI flicker when executing the
macro.

Also take care to restore the pipe_decode and wait_key settings.