neomutt: use more unique prefix char for dynamic macros
Macro definitions in neomutt must be prefix-free, so if we define a
macro named _6kgMbnD, we can no longer bind _ to a function. Use a
prefix string that is unlikely to be mapped to a key (and gives us the
notion of "macro"). Note that µ is two UTF-8 characters long, and the
limit for macro names it 8 bytes, so we have to shorten our dynamic
macros by one character.

See <https://neomutt.org/guide/configuration.html#bind-warnings>.