solarized: make sure that highlighted black is lighter than normal black
In usual terminal color schemes, the highlight colors (8-15) are lighter
than the respective normal colors (0-7). color0 is also used as the
background color in dark schemes, and color15 in light schemes.

Swap S_base03 and S_base02 to achieve that effect. Since we're using
S_base03 as background color in xterm and URxvt for solarized-dark,
highlighted black text should now also be legible.